python/requirements.txt
a2a-sdk==0.3.7
agent-framework==1.0.0b251001
agent-framework-a2a==1.0.0b251001
agent-framework-azure-ai==1.0.0b251001
agent-framework-copilotstudio==1.0.0b251001
agent-framework-core==1.0.0b251001
agent-framework-devui==1.0.0b251001
agent-framework-mem0==1.0.0b251001
agent-framework-redis==1.0.0b251001
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.11.0
asgiref==3.9.2
attrs==25.3.0
azure-ai-agents==1.2.0b5
azure-ai-projects==1.1.0b4
azure-core==1.35.1
azure-core-tracing-opentelemetry==1.0.0b12
azure-identity==1.25.0
azure-monitor-opentelemetry==1.8.1
azure-monitor-opentelemetry-exporter==1.0.0b42
azure-storage-blob==12.26.0
backoff==2.2.1
cachetools==6.2.0
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
click==8.3.0
cryptography==46.0.2
distro==1.9.0
elastic-transport==9.1.0
elasticsearch==9.1.1
fastapi==0.118.0
fixedint==0.1.6
frozenlist==1.7.0
google-api-core==2.25.1
google-auth==2.41.1
googleapis-common-protos==1.70.0
greenlet==3.2.4
grpcio==1.75.1
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
httpx-sse==0.4.1
hyperframe==6.1.0
idna==3.10
importlib_metadata==8.7.0
isodate==0.7.2
jiter==0.11.0
jsonpath-ng==1.7.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
mcp==1.15.0
mem0ai==0.1.118
microsoft-agents-activity==0.3.2
microsoft-agents-copilotstudio-client==0.3.2
microsoft-agents-hosting-core==0.3.2
ml_dtypes==0.5.3
msal==1.34.0
msal-extensions==1.3.1
msrest==0.7.1
multidict==6.6.4
numpy==2.3.3
oauthlib==3.3.1
openai==1.109.1
opentelemetry-api==1.37.0
opentelemetry-exporter-otlp-proto-common==1.37.0
opentelemetry-exporter-otlp-proto-grpc==1.37.0
opentelemetry-instrumentation==0.58b0
opentelemetry-instrumentation-asgi==0.58b0
opentelemetry-instrumentation-dbapi==0.58b0
opentelemetry-instrumentation-django==0.58b0
opentelemetry-instrumentation-fastapi==0.58b0
opentelemetry-instrumentation-flask==0.58b0
opentelemetry-instrumentation-psycopg2==0.58b0
opentelemetry-instrumentation-requests==0.58b0
opentelemetry-instrumentation-urllib==0.58b0
opentelemetry-instrumentation-urllib3==0.58b0
opentelemetry-instrumentation-wsgi==0.58b0
opentelemetry-proto==1.37.0
opentelemetry-resource-detector-azure==0.1.5
opentelemetry-sdk==1.37.0
opentelemetry-semantic-conventions==0.58b0
opentelemetry-semantic-conventions-ai==0.4.13
opentelemetry-util-http==0.58b0
packaging==25.0
ply==3.11
portalocker==3.2.0
posthog==6.7.6
propcache==0.3.2
proto-plus==1.26.1
protobuf==5.29.5
psutil==7.1.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.11.9
pydantic-settings==2.11.0
pydantic_core==2.33.2
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-multipart==0.0.20
python-ulid==3.1.0
pytz==2025.2
PyYAML==6.0.3
qdrant-client==1.15.1
redis==6.4.0
redisvl==0.9.1
referencing==0.36.2
requests==2.32.5
requests-oauthlib==2.0.0
rpds-py==0.27.1
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.43
sse-starlette==3.0.2
starlette==0.48.0
tenacity==9.1.2
tqdm==4.67.1
typing==3.7.4.3
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.37.0
uvloop==0.21.0
watchfiles==1.1.0
websockets==15.0.1
wrapt==1.17.3
yarl==1.20.1
zipp==3.23.0
